feat: Add Dockerfile and docker-entrypoint.sh
This commit adds a Dockerfile and a docker-entrypoint.sh script to the repository. The Dockerfile installs necessary dependencies, such as Python 3, AWS CLI, boto3, kubectl, and argocd. The docker-entrypoint.sh script is a simple script that executes the provided command. These additions will enable easier deployment and management of the application using Docker.